backend/requirements.txt
# requirements.txt for backend API
fastapi
uvicorn
requests
opensearch-py
Pillow
numpy
python-multipart
